network-exporter/requirements.txt
prometheus-client==0.19.0
psutil==5.9.8
speedtest-cli==2.1.3 